% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_nitrogen_deposition}
\alias{aggregate_nitrogen_deposition}
\title{Aggregates monitoring nitrogen deposition data}
\usage{
aggregate_nitrogen_deposition(data)
}
\description{
Simplifies nitrogen deposition components to broader source categories and aggregates nitrogen deposition input dataset
per year, site, ecosystem category and source category.
}
