% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_trend_results}
\alias{aggregate_trend_results}
\title{Wrangle and aggregate results from statistical meteo-normalisation of monitoring data as relative trends}
\usage{
aggregate_trend_results(trends, reference_year_fun, nmin_sites_fun)
}
\description{
Wrangle and aggregate results from statistical meteo-normalisation of monitoring data as relative trends
}
