% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality_package.R
\docType{package}
\name{airquality-package}
\alias{airquality}
\alias{airquality-package}
\title{airquality: Gather, Standardize, and Analyze Air Quality Data}
\description{
The goal of airquality is to reduce the burden of gathering, standardizing, and analyzing air quality data. This package is very much a work in progress, significant changes will occur and there is no gaurentee things will work as expected.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/B-Nilson/airquality}
  \item \url{https://b-nilson.github.io/airquality/}
  \item Report bugs at \url{https://github.com/B-Nilson/airquality/issues}
}

}
\keyword{internal}
