% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality.R
\docType{package}
\name{airquality}
\alias{airquality}
\title{Airquality Package}
\description{
The package contains the following functions:
}
\section{Current airquality}{

\emph{current_aq_df()}
}

\section{Plotting the AQI for the past two weeks}{

\emph{plot_aqi_hist()}
}

\section{Plotting Air Quality Subcomponents for the Past two Weeks}{

\emph{plot_comp_hist()}
}

\section{Current AQI of the 10 largest European Cities}{

\emph{current_aq_table()}
}

\section{Current Weather Data}{

\emph{current_weather()}
}

\section{AQI Map for Europe}{

\emph{aqi_map()}
}

\section{Shiny App}{

\emph{airquality_app()}
}

