% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{AQI}
\alias{AQI}
\title{Calculate the US AQI from pollutant observations}
\usage{
AQI(
  datetimes = Sys.time(),
  o3_8hr_ppm = NA,
  o3_1hr_ppm = NA,
  pm25_24hr_ugm3 = NA,
  pm10_24hr_ugm3 = NA,
  co_8hr_ppm = NA,
  so2_1hr_ppb = NA,
  no2_1hr_ppb = NA
)
}
\arguments{
\item{datetimes}{Vector of hourly datetimes corresponding to observations. Date gaps will be filled automatically.}

\item{o3_8hr_ppm}{(Optional) Numeric vector of hourly 1-hour mean ozone (O3) concentrations (ppm).}

\item{o3_1hr_ppm}{(Optional) Numeric vector of hourly 8-hour mean ozone (O3) concentrations (ppm).}

\item{pm25_24hr_ugm3}{(Optional) Numeric vector of hourly 24-hour mean fine particulate matter (PM2.5) concentrations (ug/m^3).}

\item{pm10_24hr_ugm3}{(Optional) Numeric vector of hourly 24-hour mean coarse particulate matter (PM10) concentrations (ug/m^3).}

\item{co_8hr_ppm}{(Optional) Numeric vector of hourly 8-hour mean carbon monoxide (CO) concentrations (ppm).}

\item{so2_1hr_ppb}{(Optional) Numeric vector of hourly 1-hour mean sulfur dioxide (SO2) concentrations (ppb).}

\item{no2_1hr_ppb}{(Optional) Numeric vector of hourly 1-hour mean nitrogen dioxide (NO2) concentrations (ppb).}
}
\value{
A tibble (data.frame) with columns:
date, AQI, risk_category, principal_pol
 and potentially more rows than `length(datetimes)` (due to any missing hours being filled with NA values).
}
\description{
Calculate the US AQI from pollutant observations
}
\examples{
AQI(o3_8hr_ppm = 0.078, o3_1hr_ppm = 0.104, pm25_24hr_ugm3 = 35.9)
AQI(o3_1hr_ppm = 0.104, pm25_24hr_ugm3 = 35.9)
}
\seealso{
Other Air Quality Standards: 
\code{\link{AQHI}()},
\code{\link{AQHI_plus}()},
\code{\link{CAAQS}()}
}
\concept{Air Quality Standards}
\concept{American (USA) Air Quality}
