% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bin_fun}
\alias{bin_fun}
\title{Return function to bin concentration depending on parameter}
\usage{
bin_fun(pollutant)
}
\description{
Return function to bin concentration depending on parameter
}
\keyword{internal}
