% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{CAAQS}
\alias{CAAQS}
\title{Assess the attainment of the Canadian Ambient Air Quality Standards (CAAQS)}
\usage{
CAAQS(
  datetimes,
  pm25_hourly = NULL,
  o3_hourly = NULL,
  no2_hourly = NULL,
  so2_hourly = NULL
)
}
\arguments{
\item{datetimes}{Vector of hourly datetime values corresponding to observations. Date gaps will be filled automatically.}

\item{pm25_hourly}{(Optional). Vector of hourly mean fine particulate matter (PM2.5) concentrations (ug/m^3).}

\item{o3_hourly}{(Optional). Vector of hourly mean ozone (O3) concentrations (ppb).}

\item{no2_hourly}{(Optional). Vector of hourly mean nitrogen dioxide (NO2) concentrations (ppb).}

\item{so2_hourly}{(Optional). Vector of hourly mean sulphur dioxide (SO2) concentrations (ppb).}
}
\value{
a list of tibbles (data.frames), one tibble per pollutant provided with annual CAAQS metrics and management levels
}
\description{
The Canadian Ambient Air Quality Standards (CAAQS) are part of a collaborative national Air Quality Management System (AQMS), to better protect human health and the environment.
Standards are defined for fine particulate matter (PM2.5), ozone (O3), nitrogen dioxide (NO2), and sulphur dioxide (SO2), and are typically updated ever 5 years.
}
\examples{
obs = data.frame(
  date = seq(lubridate::ymd_h("2024-01-01 00"),
             lubridate::ymd_h("2024-01-01 23"), "1 hours"),
  pm25 = sample(1:150, 24), o3 = sample(1:150, 24),
  no2 = sample(1:150, 24), so2 = sample(1:150, 24))
CAAQS(datetimes = obs$date, pm25_hourly = obs$pm25,
     o3_hourly = obs$o3, no2_hourly = obs$no2, so2_hourly = obs$so2)
}
\references{
\link{https://ccme.ca/en/air-quality-report}
}
\seealso{
Other Canadian Air Quality Standards: 
\code{\link{AQHI}()},
\code{\link{AQHI_plus}()}
}
\concept{Canadian Air Quality Standards}
