% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{calc_outcome}
\alias{calc_outcome}
\title{calculate specific health outcome}
\usage{
calc_outcome(conc_increment, crf, crf_conc_increment, cases)
}
\description{
calculate specific health outcome
}
\keyword{internal}
