% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{calc_rsd_nox_emission}
\alias{calc_rsd_nox_emission}
\title{RSD NOx emission in g/kg fuel}
\usage{
calc_rsd_nox_emission(NO, p, CO2, CO, HC)
}
\arguments{
\item{NO, CO2, CO, HC}{Concentrations as mixing ratios in percent.}

\item{p}{NO2 fraction of NOx.}
}
\value{
NOx emission (as NO2) in g/kg fuel.
}
\description{
RSD NOx emission in g/kg fuel
}
\keyword{internal}
