% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{calculate_all_outcomes}
\alias{calculate_all_outcomes}
\title{Calculate range of health outcomes from input-dataset (most likely value, lower and upper confidence intervals crf)}
\usage{
calculate_all_outcomes(data, conc_threshold = "lower_conc_threshold")
}
\description{
Calculate range of health outcomes from input-dataset (most likely value, lower and upper confidence intervals crf)
}
\keyword{internal}
