% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_list}
\alias{city_list}
\title{All Countries and Capitals with Latitude and Longitude Data.}
\format{
A data frame with 247 rows and 6 variables:
\describe{
\item{CountryName}{Name of the country}
\item{CapitalName}{Name of the capital}
\item{CapitalLatitude}{Capital's latitude}
\item{CapitalLongitude}{Capital's longitude}
\item{CountryCode}{Country code}
\item{ContinentName}{Continent name}

}
}
\source{
\url{http://techslides.com/list-of-countries-and-capitals}
}
\usage{
city_list
}
\description{
A dataset containing all countries' capitals including country code and coordinates.
}
\keyword{datasets}
