% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{classify}
\alias{classify}
\title{Classify values into labelled classes (lower bound of the first class included)}
\usage{
classify(x, breaks, labels)
}
\arguments{
\item{x}{Numeric vector.}

\item{breaks}{Class boundaries.}

\item{labels}{Class labels.}
}
\value{
Factor with \code{labels} as levels.
}
\description{
Classify values into labelled classes (lower bound of the first class included)
}
\keyword{internal}
