% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{classify_estimated}
\alias{classify_estimated}
\title{Class of the estimated (modelled) part of the nitrogen deposition}
\usage{
classify_estimated(
  estimated,
  breaks = c(0, 5, 12, Inf),
  labels = c("<5 kg-N", "5-12 kg-N", ">12 kg-N")
)
}
\arguments{
\item{estimated}{Estimated deposition in kg N/ha/a.}

\item{breaks, labels}{Class boundaries and labels.}
}
\value{
Factor "<5 kg-N", "5-12 kg-N", ">12 kg-N".
}
\description{
Class of the estimated (modelled) part of the nitrogen deposition
}
\keyword{internal}
