% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert between common units}
\usage{
convert_units(x, in_unit, out_unit, keep_units = FALSE)
}
\arguments{
\item{x}{Vector of numeric values to be converted.}

\item{in_unit}{A single character value indicating the units of `x`. See `units::available_units()` for options.}

\item{out_unit}{A single character value indicating the units to convert `x` to. See `units::available_units()` for options.}

\item{keep_units}{(Optional) A single logical (TRUE or FALSE) value indicating if the units of `x` should be kept.}
}
\value{
a vector of `x` converted from `in_unit` to `out_unit`
}
\description{
`convert_units` provides a simple way to convert between units, leveraging the `units` package.
}
\examples{
convert_units(c(-20:20), in_unit = "degC", out_unit = "degF")
convert_units(c(0:10), in_unit = "ppm", out_unit = "ppb", keep_units = TRUE)
}
\concept{Utilities}
