% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{current_aq_table}
\alias{current_aq_table}
\title{Current Airquality Index of Top 10 Major European Cities}
\usage{
current_aq_table()
}
\description{
This function produces a table with the current airquality index (AQI) of the top 10 major European cities.
}
\examples{
current_aq_table


}
