% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{current_weather}
\alias{current_weather}
\title{Current Weather Data}
\usage{
current_weather(city = "Amsterdam", country = "NL")
}
\arguments{
\item{city}{City (in quotation marks)}

\item{country}{Country Code (according to ISO 3166; e.g.: NL for Netherlands; in quotation marks)}
}
\value{
Current weather data.
}
\description{
With this function you can get the current weather data of any town.
}
\examples{
current_weather("Amsterdam", "NL")


}
