% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposition.R
\name{exposition_collections}
\alias{exposition_collections}
\title{geo.admin collections used for the population exposition}
\usage{
exposition_collections
}
\description{
geo.admin collections used for the population exposition
}
\keyword{internal}
