% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposition.R
\name{exposition_parameters}
\alias{exposition_parameters}
\title{Parameter names of the cell-table pollutant columns, as used in all outputs}
\usage{
exposition_parameters
}
\description{
Parameter names of the cell-table pollutant columns, as used in all outputs
}
\keyword{internal}
