% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposition.R
\name{expositionpars}
\alias{expositionpars}
\title{Parameter-specific settings of the exposition plots}
\usage{
expositionpars(parameter)
}
\arguments{
\item{parameter}{Parameter, e.g. "NO2", or "Ndep".}
}
\value{
List with \code{barwidth}, \code{xbreaks} (their range is the x range), \code{aggregation} and \code{metric}.
}
\description{
Parameter-specific settings of the exposition plots
}
\keyword{internal}
