% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_threshold}
\alias{extract_threshold}
\title{Extract target threshold values from threshold table for plotting}
\usage{
extract_threshold(
  threshold_values,
  pollutant = NULL,
  metric = "Jahresmittel",
  interval = "y1",
  unit = "µg/m3",
  source = c("LRV Grenzwert", "WHO Richtwert")
)
}
\description{
Extract target threshold values from threshold table for plotting
}
\keyword{internal}
