% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_ressources}
\alias{filter_ressources}
\title{Get entry from ressources.csv by internal id}
\usage{
filter_ressources(ressources, internal_id)
}
\description{
Get entry from ressources.csv by internal id
}
\keyword{internal}
