% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{geocoding}
\alias{geocoding}
\title{Geocoding}
\usage{
geocoding(city, country)
}
\arguments{
\item{city}{City (in quotation marks)}

\item{country}{Country Code (according to ISO 3166; e.g.: NL for Netherlands; in quotation marks)}
}
\value{
Coordinates (latitude, longitude)
}
\description{
This function gives the coordinates of a given city. It used for further functions which require coordinates instead of city names.
}
\details{
It is important to enter the city and the country.
}
\examples{
geocoding("Amsterdam", "NL")

}
