% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abgov_stations.R
\name{get_abgov_stations}
\alias{get_abgov_stations}
\title{Download air quality station metadata from the Alberta (Canada) Government}
\usage{
get_abgov_stations(..., use_sf = FALSE)
}
\arguments{
\item{...}{Not used. For compatibility with other metadata functions and future expansion.}

\item{use_sf}{(Optional) a single logical (TRUE/FALSE) value indicating whether or not to return a spatial object. using the `sf` package}
}
\value{
A tibble of metadata for Alberta air quality monitoring stations.
}
\description{
Air pollution monitoring in Canada is done by individual Provinces/Territories,
primarily as a part of the federal National Air Pollution Surveillance (NAPS) program.
The Province of Alberta hosts it's hourly air quality metadata
through a public API.

[get_abgov_stations()] provides an easy way to retrieve this metadata (typically to determine station names to pass to `get_abgov_data()`)
}
\examples{
\donttest{
# Normal usage
get_abgov_stations()
# if spatial object required
get_abgov_stations(use_sf = TRUE)
}
}
\seealso{
[get_abgov_data()]

Other Data Collection: 
\code{\link{get_abgov_data}()},
\code{\link{get_airnow_data}()},
\code{\link{get_airnow_stations}()},
\code{\link{get_bcgov_data}()},
\code{\link{get_bcgov_stations}()},
\code{\link{purpleair_api}()}

Other Canadian Air Quality: 
\code{\link{CAAQS}()},
\code{\link{get_abgov_data}()},
\code{\link{get_bcgov_data}()},
\code{\link{get_bcgov_stations}()},
\code{\link{purpleair_api}()}
}
\concept{Canadian Air Quality}
\concept{Data Collection}
