% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection.R
\name{get_airnow_data}
\alias{get_airnow_data}
\title{Download air quality station observations from the US EPA "AirNow" platform}
\usage{
get_airnow_data(stations = "all", date_range, raw = FALSE)
}
\arguments{
\item{stations}{(Optional) Either "all" or a character vector specifying AQS IDs for stations to filter data to.
If "all" not provided, data for all stations for each hour in `date_range` are still downloaded,
but only data for desired stations is returned. Default is "all".}

\item{date_range}{A datetime vector (or a character vector with UTC dates in "YYYY-MM-DD HH" format) with either 1 or 2 values.
Providing a single value will return data for that hour only,
whereas two values will return data between (and including) those times}

\item{raw}{(Optional) A single logical (TRUE or FALSE) value indicating if
raw data files desired (i.e. without standardized column names). Default is FALSE.}
}
\value{
A tibble of hourly observation data for desired station(s) and date range where available.
Columns available will vary depending on available data from station(s).
}
\description{
A short description...
}
\examples{
# Get data for all stations for first 3 hours of Jan 2019
date_range = lubridate::ymd_h(c("2019-01-01 00", "2019-01-01 02"), tz = "Etc/GMT+8")
get_airnow_data("all", date_range)

# Get data for two specific stations for first 3 hours of Jan 2019
date_range = lubridate::ymd_h(c("2019-01-01 00", "2019-01-01 02"), tz = "Etc/GMT+8")
get_airnow_data(c("000010102", "000010401"), date_range)

# Get non-standardized data for all stations for first 3 hours of Jan 2019
date_range = lubridate::ymd_h(c("2019-01-01 00", "2019-01-01 02"), tz = "Etc/GMT+8")
get_airnow_data("all", date_range, raw = TRUE)
}
\seealso{
Other Data Collection: 
\code{\link{get_bc_data}()}

Other USA Air Quality: 
\code{\link{AQI}()}
}
\concept{Data Collection}
\concept{USA Air Quality}
