% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_airnow.R
\name{get_airnow_stations}
\alias{get_airnow_stations}
\title{Download air quality station metadata from the US EPA "AirNow" platform}
\usage{
get_airnow_stations(dates = Sys.time(), use_sf = FALSE)
}
\arguments{
\item{dates}{(Optional) one or more date values indicating the day(s) to get metadata for.
Default is the current date.}

\item{use_sf}{(Optional) a single logical (TRUE/FALSE) value indicating whether or not to return a spatial object. using the `sf` package}
}
\value{
A tibble of metadata for the air quality monitoring stations on AirNow.
}
\description{
AirNow is a US EPA nationwide voluntary program which hosts non-validated air quality
observation data from stations in the US and many other countries globally.

The AirNow API provides access to daily metadata files for the available stations at that time.

[get_airnow_stations()] provides an easy way to retrieve this metadata (typically to determine station id's to pass to `get_airnow_data()`)
}
\examples{
\donttest{
# Normal usage
get_airnow_stations()
# if spatial object required
get_airnow_stations(use_sf = TRUE)
# if data for past/specific years required
get_airnow_stations(dates = lubridate::ymd("2022-01-01"))
}
}
\seealso{
[get_airnow_data()]

Other Data Collection: 
\code{\link{get_abgov_data}()},
\code{\link{get_abgov_stations}()},
\code{\link{get_airnow_data}()},
\code{\link{get_bcgov_data}()},
\code{\link{get_bcgov_stations}()},
\code{\link{purpleair_api}()}

Other USA Air Quality: 
\code{\link{AQI}()},
\code{\link{get_airnow_data}()}
}
\concept{Data Collection}
\concept{USA Air Quality}
