% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{get_base_scenario_year}
\alias{get_base_scenario_year}
\title{Get year of health-outcome base scenario: either provided year or a provided function}
\usage{
get_base_scenario_year(base = "min", ...)
}
\description{
Get year of health-outcome base scenario: either provided year or a provided function
}
\keyword{internal}
