% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection.R
\name{get_bc_data}
\alias{get_bc_data}
\title{Download air quality station observations from the British Columbia (Canada) Government}
\usage{
get_bc_data(stations, date_range, raw = FALSE)
}
\arguments{
\item{stations}{A character vector of one or more station IDs (BC EMS IDs) identifying stations data desired for. See also: get_bc_stations()}

\item{date_range}{A datetime vector (or a character vector with UTC dates in "YYYY-MM-DD HH" format) with either 1 or 2 values.
Providing a single value will return data for that hour only,
whereas two values will return data between (and including) those times.
Interpreted as "backward-looking", so a date of "2019-01-01 01" covers from "2019-01-01 00:01"- "2019-01-01 01:00".}

\item{raw}{(Optional) A single logical (TRUE or FALSE) value indicating if raw data files desired (i.e. without standardized column names). Default is FALSE.}
}
\value{
A tibble of hourly observation data for desired station(s) and date range where available.
Columns available will vary depending on available data from station(s).
}
\description{
Air pollution monitoring is done by individial Provinces/Territories in Canada.
The Province of British Columbia provides hourly observations (both QA/QC'ed and real-time raw data) through a public FTP site.

Annual QA/QC'ed files are available for each monitoring station, usually 1-2 years out of date due to the QA/QC process.
A single file is available for each station for all data following the QA/QC'ed years
(with potentially 2+ years of data depending on the time since the last QA/QC'ed dataset was created).

`get_bc_data()` provides an easy way to retrieve data for stations managed by the BC Government (using BC's "EMS IDs" - see `get_bc_stations()`) and a specified date or date range.

Due to the FTP site's file structure, data retrieval time is proportional to the number of stations requested
as well as the number of years (PST timezone) desired, with potentially longer retrieval times
for non-QA/QC'ed years depending on the number of years of data in the stations raw data file.
}
\examples{
# For a single station
station = "0450307" # EMS IDs - see get_bc_stations()
# For the years 2019 and 2020
date_range = lubridate::ymd_h(c("2019-01-01 00", "2020-12-31 23"), tz = "Etc/GMT+8")
get_bc_data(station, date_range)

# For multiple stations
stations = c("0450307", "E206898") # EMS IDs - see get_bc_stations()
# For first week of January 2019
date_range = lubridate::ymd_h(c("2019-01-01 00", "2019-01-07 23"), tz = "Etc/GMT+8")
get_bc_data(stations, date_range)
}
\seealso{
[get_bc_stations()]

Other Data Collection: 
\code{\link{get_airnow_data}()}

Other Canadian Air Quality: 
\code{\link{AQHI}()},
\code{\link{AQHI_plus}()},
\code{\link{CAAQS}()}
}
\concept{Canadian Air Quality}
\concept{Data Collection}
