% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgov_data.R
\name{get_bcgov_data}
\alias{get_bcgov_data}
\title{Download air quality station observations from the British Columbia (Canada) Government}
\usage{
get_bcgov_data(stations, date_range, raw = FALSE, verbose = TRUE)
}
\arguments{
\item{stations}{A character vector of one or more station IDs (BC EMS IDs) to try and get data desired for (see [get_bcgov_stations()]).}

\item{date_range}{A datetime vector (or a character vector with UTC dates in "YYYY-MM-DD HH" format) with either 1 or 2 values.
Providing a single value will return data for that hour only,
whereas two values will return data between (and including) those times.
Dates are "backward-looking", so a value of "2019-01-01 01:00" covers from "2019-01-01 00:01"- "2019-01-01 01:00".}

\item{raw}{(Optional) A single logical (TRUE or FALSE) value indicating
if raw data files desired (i.e. without a standardized format). Default is FALSE.}

\item{verbose}{(Optional) A single logical (TRUE or FALSE) value indicating if
non-critical messages/warnings should be printed}
}
\value{
A tibble of hourly observation data for desired station(s) and date range where available.
Columns returned will vary depending on available data from station(s).

Dates UTC time and are "backward-looking", so a value of "2019-01-01 01:00" covers from "2019-01-01 00:01"- "2019-01-01 01:00".
}
\description{
Air pollution monitoring in Canada is done by individual Provinces/Territories,
primarily as a part of the federal National Air Pollution Surveillance (NAPS) program.
The Province of British Columbia hosts it's hourly air quality observations
through a public FTP site, providing both historic QA/QC'ed and real-time raw data.

Annual QA/QC'ed files are available on the BC FTP site for each monitoring station,
however these are usually 1-2 years out of date due to the QA/QC process.
A single file is available for each station for all non-QA/QC'ed years, which has
potentially 0-2+ years of data depending on the time since the last QA/QC'ed dataset was created).

[get_bcgov_data()] provides an easy way to retrieve these observations using
BC's station "EMS IDs"  (see [get_bcgov_stations()]) and a specified date or date range.

Due to the FTP site's file structure, data retrieval time is proportional to the number of stations requested
as well as the number of years of data (PST timezone) desired. There can be potentially longer retrieval times
for non-QA/QC'ed years depending on the current time since the last QA/QC'ed year due to larger file sizes.
}
\examples{
# For a single station
station = "0450307" # EMS IDs - see get_bcgov_stations()
# For the years 2019 and 2020
date_range = lubridate::ymd_h(c("2019-01-01 00", "2020-12-31 23"), tz = "Etc/GMT+8")
get_bcgov_data(station, date_range)

# For multiple stations
stations = c("0450307", "E206898") # EMS IDs - see get_bcgov_stations()
# For first week of January 2019
date_range = lubridate::ymd_h(c("2019-01-01 00", "2019-01-07 23"), tz = "Etc/GMT+8")
get_bcgov_data(stations, date_range)
}
\seealso{
[get_bcgov_stations()]

Other Data Collection: 
\code{\link{get_airnow_data}()},
\code{\link{purpleair_api}()}

Other Canadian Air Quality: 
\code{\link{AQHI}()},
\code{\link{AQHI_plus}()},
\code{\link{CAAQS}()},
\code{\link{purpleair_api}()}
}
\concept{Canadian Air Quality}
\concept{Data Collection}
