% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{get_bfs_asset_url}
\alias{get_bfs_asset_url}
\title{Get the download url of a BFS asset from the BFS DAM api}
\usage{
get_bfs_asset_url(bfs_nr)
}
\arguments{
\item{bfs_nr}{BFS order number, e.g. "px-x-0102020300_101".}
}
\value{
Url of the master file of the asset.
}
\description{
Get the download url of a BFS asset from the BFS DAM api
}
\keyword{internal}
