% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thingspeak_data.R
\name{get_thingspeak_data}
\alias{get_thingspeak_data}
\title{Get data from a ThingSpeak channel}
\usage{
get_thingspeak_data(
  channel_id,
  read_params = NULL,
  ...,
  try_to_include_units = TRUE
)
}
\arguments{
\item{channel_id}{ThingSpeak channel ID}

\item{read_params}{(Optional) parameters to include in the request created using [ThingSpeakReadParams()].
See \href{https://www.mathworks.com/help/thingspeak/readdata.html}{the thingspeak documentation} for valid parameters or [ThingSpeakReadParams()].
Defaults to NULL, which uses `...` to build `read_params` instead.}

\item{...}{(Optional) Named parameters to include in the request if read_params is NULL.
See \href{https://www.mathworks.com/help/thingspeak/readdata.html}{the thingspeak documentation} for valid parameters or [ThingSpeakReadParams()].
Will be used to build `read_params` using [ThingSpeakReadParams()].
Default (nothing passed) is equivalent to `read_params = ThingSpeakReadParams()`.}

\item{try_to_include_units}{If TRUE, the function will try to include units in the response by looking for "\[unit\]" in the column names.}
}
\value{
A list with channel metadata and a tibble containing the data from the ThingSpeak channel.
}
\description{
Get data from a ThingSpeak channel
}
\examples{
\dontrun{
data <- get_thingspeak_data(channel_id = 123456)
}
}
