% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_timezone}
\alias{get_timezone}
\title{Wrapper for looking up timezone of locations from lat/lng coords}
\usage{
get_timezone(lng, lat, method = "accurate", ...)
}
\arguments{
\item{lng}{Vector of numeric values indicating location longitudes (decimal degrees) to lookup.}

\item{lat}{Vector of numeric values indicating location latitudes (decimal degrees) to lookup.}

\item{method}{A single character value indicating the lookup method to use (see `?lutz::tz_lookup_coords`)}

\item{...}{(Optional) Additional paramaters to pass to `lutz::tz_lookup_coords`}
}
\value{
a character vector with the same length as `lat` and `lng` indicating the locations likely timezone.
}
\description{
`get_timezone` is a wrapper for the `lutz::tz_lookup_coords` function. See `?lutz::tz_lookup_coords` for more details.
}
\examples{
get_timezone(-105.053144, 69.116178)
}
\seealso{
Other Utilities: 
\code{\link{convert_units}()},
\code{\link{swap}()}
}
\concept{Utilities}
