% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_emissions}
\alias{ggplot_emissions}
\title{Plot emission budget data employing structured coloring using ggplot2 ... this is not ideal, but the best I can do}
\usage{
ggplot_emissions(
  data,
  relative = FALSE,
  pos = "stack",
  width = 0.8,
  theme = ggplot2::theme_minimal()
)
}
\description{
Plot emission budget data employing structured coloring using ggplot2 ... this is not ideal, but the best I can do
}
\keyword{internal}
