% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_expo_hist}
\alias{ggplot_expo_hist}
\title{Plot exposition distribution histogram using ggplot2}
\usage{
ggplot_expo_hist(
  data,
  x,
  y,
  barwidth = 1,
  xlims = c(0, NA),
  xbreaks = waiver(),
  titlelab = NULL,
  captionlab = NULL,
  xlabel = NULL,
  threshold = list(value = NA, label = NULL, labelsize = 4, linetype = 2, linesize = 1),
  fill_scale = NULL,
  theme = ggplot2::theme_minimal()
)
}
\description{
Plot exposition distribution histogram using ggplot2
}
\keyword{internal}
