% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_helpers.R
\name{groups_emission_subsector}
\alias{groups_emission_subsector}
\title{Merge groups with small emissions in emission budget datasets}
\usage{
groups_emission_subsector(data, threshold_fraction = 0.05, first = 1:2)
}
\description{
Merge groups with small emissions in emission budget datasets
}
\keyword{internal}
