% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{highslows}
\alias{highslows}
\title{Collect Highs and Lows for PM2.5 per day}
\usage{
highslows(data)
}
\arguments{
\item{data}{a dataframe of hourly PurpleAir sensor data from the CEHAT website}
}
\value{
a dataframe of the highs and lows of PurpleAir sensor data, organized by sensor
}
\description{
This function collects the daily high and low values of PM2.5 for each sensor into a dataframe.
It assumes that the data has the columns 'timestamp' and 'longitude'.
It also uses an EPA correction factor for PurpleAir sensors.
Anyidentical readings of duplicate type will be dropped
}
