% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{history_aq}
\alias{history_aq}
\title{co concentration of the last two weeks}
\usage{
history_aq(city, country)
}
\arguments{
\item{city}{City (in quotation marks)}

\item{country}{Country Code (according to ISO 3166; e.g.: NL for Netherlands; in quotation marks)}
}
\value{
Line graph of the co concentration of the last two weeks.
}
\description{
This function plots the co concentration of the last two weeks of any town.
}
\details{
The x-axis does not display the correct time yet.
}
\examples{
history_aq("Amsterdam", "NL")

}
