% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{history_aq_list}
\alias{history_aq_list}
\title{Two Week Airquality Data}
\usage{
history_aq_list(city, country)
}
\arguments{
\item{city}{City (in quotation marks)}

\item{country}{Country Code (according to ISO 3166; e.g.: NL for Netherlands; in quotation marks)}
}
\value{
The airquality index (aqi), the subcomponents (co, no, no2, o3, so2, pm2_5, pm10, nh3), and the time stamps in a list object.
}
\description{
With this function you can get the air quality data of any town for the last two weeks.
}
\details{
The airquality index ranges from 1 = Good to 5 = Very Poor.
}
\examples{
history_aq_list("Amsterdam", "NL")

}
