% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{is_model_year}
\alias{is_model_year}
\title{Does a \code{vehicle_euronorm} entry of the RSD metadata hold a vehicle model year?}
\usage{
is_model_year(x)
}
\arguments{
\item{x}{Character vector.}
}
\description{
Does a \code{vehicle_euronorm} entry of the RSD metadata hold a vehicle model year?
}
\keyword{internal}
