% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{label_big_mark}
\alias{label_big_mark}
\title{Axis labels with an apostrophe as thousands separator}
\usage{
label_big_mark(x)
}
\arguments{
\item{x}{Numbers.}
}
\value{
Character vector, e.g. "25'000".
}
\description{
Axis labels with an apostrophe as thousands separator
}
\keyword{internal}
