% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends.R
\name{label_ozone_metric}
\alias{label_ozone_metric}
\title{Add the metric to the name of ozone, which has two parameters}
\usage{
label_ozone_metric(data)
}
\arguments{
\item{data}{Trend data with \code{pollutant} and \code{parameter}.}
}
\value{
\code{data} with \code{pollutant} "Ozon, \if{html}{\out{<metric>}}" for ozone.
}
\description{
Add the metric to the name of ozone, which has two parameters
}
\keyword{internal}
