% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposition.R
\name{ndep_collection}
\alias{ndep_collection}
\title{geo.admin collection of the critical load exceedance for nitrogen}
\usage{
ndep_collection
}
\description{
geo.admin collection of the critical load exceedance for nitrogen
}
\keyword{internal}
