% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{parameter_setting}
\alias{parameter_setting}
\title{Setting of one parameter from a list of settings per parameter}
\usage{
parameter_setting(settings, parameter)
}
\arguments{
\item{settings}{Named list, one element per parameter (e.g. \code{plot_axes_timeseries} of
\verb{scripts/_plot_setup.R}).}

\item{parameter}{Parameter.}
}
\value{
The element of \code{parameter}. Stops with an error of class \code{airquality_plot_error} if there is
none, naming the parameters with a setting.
}
\description{
Setting of one parameter from a list of settings per parameter
}
\keyword{internal}
