% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_all_expo_cumul_ndep}
\alias{plot_all_expo_cumul_ndep}
\title{Wrapper to plot sensitive ecosystem nitrogen deposition cumulative exposition distribution in a list by year}
\usage{
plot_all_expo_cumul_ndep(data, threshold_ndep, sub = "im Kanton Zürich")
}
\description{
Wrapper to plot sensitive ecosystem nitrogen deposition cumulative exposition distribution in a list by year
}
\keyword{internal}
