% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_aqi_hist}
\alias{plot_aqi_hist}
\title{Airquality Index Plot (past two weeks)}
\usage{
plot_aqi_hist(city, country)
}
\arguments{
\item{city}{City (in quotation marks)}

\item{country}{Country Code (according to ISO 3166; e.g.: NL for Netherlands; in quotation marks)}
}
\value{
A plot of the airquality index over the past two weeks.
}
\description{
With this function you can plot  the airquality index of the last two weeks.
}
\details{

}
\examples{
plot_comp_hist("Amsterdam", "NL")

}
