% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_comp_hist}
\alias{plot_comp_hist}
\title{Airquality Subcomponents Plot (past two weeks)}
\usage{
plot_comp_hist(city = "Amsterdam", country = "NL", component = "co")
}
\arguments{
\item{city}{City (in quotation marks)}

\item{country}{Country Code (according to ISO 3166; e.g.: NL for Netherlands; in quotation marks)}

\item{component}{The subcomponent you wish to plot}
}
\value{
A plot of the airquality subcomponent over the past two weeks.
}
\description{
With this function you can plot each of the subcomponents for the airquality of the last two weeks.
}
\details{

}
\examples{
plot_comp_hist("Amsterdam", "NL", "co")

}
