% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends.R
\name{plot_emission_trends_relative}
\alias{plot_emission_trends_relative}
\title{Plot the relative emissions of all pollutants vs. the reference year}
\usage{
plot_emission_trends_relative(
  data,
  reference_year,
  linewidth = 1,
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{data}{Data as returned by \code{\link[=emission_trends_relative]{emission_trends_relative()}}.}

\item{reference_year}{Reference year (subtitle).}

\item{linewidth}{Width of the lines.}

\item{theme}{ggplot2 theme.}
}
\value{
A ggplot object.
}
\description{
Plot the relative emissions of all pollutants vs. the reference year
}
\keyword{internal}
