% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposition.R
\name{plot_pars_popmean_timeseries}
\alias{plot_pars_popmean_timeseries}
\title{Plot the population-weighted mean per year and its reduction vs. the base year, one plot per parameter}
\usage{
plot_pars_popmean_timeseries(
  data,
  parameters,
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{data}{\code{data_exposition_weighted_means_canton.csv}.}

\item{parameters}{Parameters to plot.}

\item{theme}{ggplot2 theme.}
}
\value{
Named list of ggplot objects, one per parameter.
}
\description{
Plot the population-weighted mean per year and its reduction vs. the base year, one plot per parameter
}
\keyword{internal}
