% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emissions.R
\name{plot_rsd_per_yearmeas}
\alias{plot_rsd_per_yearmeas}
\title{Plot the trend of the RSD NOx emissions per year of measurement and fuel type}
\usage{
plot_rsd_per_yearmeas(data, theme = ggplot2::theme_minimal())
}
\arguments{
\item{data}{\code{data_nox_emissions_rsd_per_yearmeas.csv}.}

\item{theme}{ggplot2 theme.}
}
\value{
A ggplot object (loess smooth with 95 \% confidence band).
}
\description{
Plot the trend of the RSD NOx emissions per year of measurement and fuel type
}
\keyword{internal}
