% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries_ndep_bars}
\alias{plot_timeseries_ndep_bars}
\title{Plot yearly nitrogen deposition timeseries using ggplot2}
\usage{
plot_timeseries_ndep_bars(
  data,
  xlim = NULL,
  xbreaks = waiver(),
  linewidth = 1,
  color = "red3",
  title = "Luftqualitätsmesswerte - Stickstoffeintrag in empfindliche Ökosysteme"
)
}
\description{
Plot yearly nitrogen deposition timeseries using ggplot2
}
\keyword{internal}
