% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries_trend_relative}
\alias{plot_timeseries_trend_relative}
\title{Plot pre-compiled relative trends of emissions and immissions vs. a reference year for various pollutants}
\usage{
plot_timeseries_trend_relative(
  data_trends,
  detailed = FALSE,
  pt_size = 1.5,
  linewdth = 1,
  facet_ncol = NULL,
  facet_scale = "free_y",
  theme = ggplot2::theme_minimal(),
  titlelab = NULL,
  captionlab = NULL
)
}
\description{
Plot pre-compiled relative trends of emissions and immissions vs. a reference year for various pollutants
}
\keyword{internal}
