% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_data_trends}
\alias{prepare_data_trends}
\title{Wrangle data from air quality monitoring and meteoswiss (d1) for statistical meteo-normalisation analysis}
\usage{
prepare_data_trends(data_aq, data_met)
}
\description{
Wrangle data from air quality monitoring and meteoswiss (d1) for statistical meteo-normalisation analysis
}
