% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_life_expectancy_data}
\alias{prepare_life_expectancy_data}
\title{Wrangle life-expectancy data from *.px format into tibble()}
\usage{
prepare_life_expectancy_data(data)
}
\description{
Wrangle life-expectancy data from *.px format into tibble()
}
\keyword{internal}
