% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{prepare_ndep_site_meta}
\alias{prepare_ndep_site_meta}
\title{Classify the nitrogen deposition sites}
\usage{
prepare_ndep_site_meta(site_meta)
}
\arguments{
\item{site_meta}{Site metadata (\code{airquality.data::site_meta_ndep}).}
}
\value{
One row per site and ecosystem with \code{site}, \code{ecosys} (recoded), \code{x}, \code{y}, \code{masl},
\code{siteclass}, \code{emissionclass}, \code{gve_5km}, \code{nh3_emission_1km}, \code{cln} (critical load).
}
\description{
Classify the nitrogen deposition sites
}
\keyword{internal}
