% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{prepare_population_by_age}
\alias{prepare_population_by_age}
\title{Prepare the population per year, sex and age of the canton}
\usage{
prepare_population_by_age(data)
}
\arguments{
\item{data}{Population per municipality (and further groups) with \code{jahr}, \code{geschlecht},
\code{einjahresaltersklasse} and \code{anzahl}.}
}
\value{
Tibble with \code{year}, \code{sex} ("male", "female"), \code{age} and \code{population}.
}
\description{
Prepare the population per year, sex and age of the canton
}
\keyword{internal}
