% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair_data.R
\name{purpleair_api}
\alias{purpleair_api}
\title{Interface with the PurpleAir API}
\usage{
purpleair_api(
  read_key = NULL,
  write_key = NULL,
  channel,
  parameters = NULL,
  verbose = TRUE
)
}
\arguments{
\item{read_key}{(Optional) A single character value of your PurpleAir API read key (see develop.purpleair.com). If NULL, write_key must be provided.}

\item{write_key}{(Optional) A single character value of your PurpleAir API write key (see develop.purpleair.com). If NULL, read_key must be provided.}

\item{channel}{A single character value indicating what API channel to use (Currently supported: keys, organization, sensors, groups)}

\item{parameters}{(Optional) A named list containing paramaters to be supplied to the API request (see api.purpleair.com).}

\item{verbose}{(Optional) A single logical (TRUE or FALSE) value indicating if
non-critical messages/warnings should be printed.}
}
\value{
If channel either "keys" or "organization", or invalid/missing parameters, then a list with the call results
If sensors or groups channel, then a tibble with the data returned from the request
}
\description{
Your Description Here
}
\examples{
# Written to .Renviron - make your own here: https://develop.purpleair.com/
# and run usethis::edit_r_environ("project") to open the environ file for editting
# Entries should look like `purpleair_api_read = "YOUR-API-KEY-HERE"`
\donttest{
read_key = Sys.getenv("purpleair_api_read")
write_key = Sys.getenv("purpleair_api_write")

parameters = list( # see api.purpleair.com for more
  nwlat = 63.595851, nwlng = -135.899856, 
  selat = 63.592657, selng = -135.891057,
  fields = "temperature", 
  sensor_index = 198385, 
  start_timestamp = Sys.time() |> 
    lubridate::with_tz("UTC") - lubridate::minutes(15))


# Get Sensors Data
test = purpleair_api(read_key = read_key, channel = "sensors", 
  parameters = parameters[1:5])

# Get Sensor Data
test = purpleair_api(read_key = read_key, channel = "sensors", 
  parameters = parameters[5:6])

# Get Sensor History
test = purpleair_api(read_key = read_key, channel = "sensors", 
 parameters = parameters[5:7])
}
}
\seealso{
Other Data Collection: 
\code{\link{get_airnow_data}()},
\code{\link{get_bcgov_data}()}

Other Canadian Air Quality: 
\code{\link{AQHI}()},
\code{\link{AQHI_plus}()},
\code{\link{CAAQS}()},
\code{\link{get_bcgov_data}()}
}
\concept{Canadian Air Quality}
\concept{Data Collection}
\concept{USA Air Quality
#'}
