% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bfs_life_expectancy_data}
\alias{read_bfs_life_expectancy_data}
\title{Reads Swiss BFS life expectancy data ("Kohortensterbetafeln) from official api}
\usage{
read_bfs_life_expectancy_data(destination_path = tempdir())
}
\description{
Reads Swiss BFS life expectancy data ("Kohortensterbetafeln) from official api
}
