% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends.R
\name{recode_trend_types}
\alias{recode_trend_types}
\title{Label the types of the aggregated trends as in the legend}
\usage{
recode_trend_types(data)
}
\arguments{
\item{data}{\code{data_airquality_trends_relative_aggregated_y1.csv}.}
}
\value{
\code{data} with \code{type} "Median Trend", "Median Messwerte" or "Emission".
}
\description{
Label the types of the aggregated trends as in the legend
}
\keyword{internal}
