% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{restricted_file}
\alias{restricted_file}
\title{Path of a restricted (non-public) input file}
\usage{
restricted_file(path)
}
\arguments{
\item{path}{Path of the file in \verb{data/restricted/}.}
}
\value{
\code{path}, if the file exists. Stops with an error of class \code{airquality_input_error} otherwise,
pointing to \code{data/restricted/README.md}.
}
\description{
Path of a restricted (non-public) input file
}
\keyword{internal}
