% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rf_meteo_normalisation}
\alias{rf_meteo_normalisation}
\title{Do site- and parameter-specific statistical meteo-normalisation based on package 'rmweather' from pre-compiled monitoring data in order to derive the meteo-normalised time-trend}
\usage{
rf_meteo_normalisation(
  data,
  trend_vars,
  frac_train = 0.8,
  ntrees = 300,
  nsamples = 300,
  verbose = TRUE,
  minimal = TRUE,
  coverage = 0.8
)
}
\description{
Do site- and parameter-specific statistical meteo-normalisation based on package 'rmweather' from pre-compiled monitoring data in order to derive the meteo-normalised time-trend
}
