% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{rsd_meta_wide}
\alias{rsd_meta_wide}
\title{Vehicle metadata in wide format: one column per metadata parameter}
\usage{
rsd_meta_wide(meta)
}
\arguments{
\item{meta}{Vehicle metadata in long format (see \code{\link[=prepare_rsd]{prepare_rsd()}}).}
}
\description{
Vehicle metadata in wide format: one column per metadata parameter
}
\keyword{internal}
