% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_helpers.R
\name{simplify_nitrogen_parameters}
\alias{simplify_nitrogen_parameters}
\title{Merge components of nitrogen deposition to broader source categories in nitrogen deposition dataset}
\usage{
simplify_nitrogen_parameters(data)
}
\description{
Merge components of nitrogen deposition to broader source categories in nitrogen deposition dataset
}
\keyword{internal}
