% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{southgate}
\alias{southgate}
\title{Add a polygon map of South Gate}
\usage{
southgate()
}
\value{
an easily plot-able polygon of South Gate, of the class Spatial Polygons DF
}
\description{
This function produces a line drawing of the city of South Gate, classed as a Spatial Polygons Data Frame
}
