% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{southgate_grid}
\alias{southgate_grid}
\title{Create a grid map of South Gate}
\usage{
southgate_grid()
}
\value{
an easily plot-able grid of South Gate, of the class Spatial Pixels DF; useful for kriging
}
\description{
This function produces a grid of the city of South Gate, classed as a Spatial Pixels Data Frame
}
