% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap}
\alias{swap}
\title{Swap out values in a vector}
\usage{
swap(x, what, with)
}
\arguments{
\item{x}{Vector of values to be have certain values swapped out.}

\item{what}{One or more values to be replaced with `with` throughout `x`.}

\item{with}{A single value to replace `what` for throughout `x`.}
}
\value{
a vector of `x` where all instances of `what` are replaced with `with`
}
\description{
`swap` provides a simple way to switch out certain values in a vector. It is useful for replacing NA's, Infinites, and erroneous values.
}
\examples{
swap(c(-20:20, NA), what = NA, with = -1)
swap(c(-20:20, Inf), what = Inf, with = NA)
swap(c(-20:20), what = Inf, with = NA)
}
\seealso{
Other Utilities: 
\code{\link{convert_units}()},
\code{\link{get_timezone}()},
\code{\link{on_error}()}
}
\concept{Utilities}
