% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends.R
\name{trend_data_overview}
\alias{trend_data_overview}
\title{Trend data of the overview: emissions, median trend and median of the measurements}
\usage{
trend_data_overview(data_agg)
}
\arguments{
\item{data_agg}{\code{data_airquality_trends_relative_aggregated_y1.csv}.}
}
\value{
\code{data_agg} filtered and labelled, \code{type} as factor.
}
\description{
Trend data of the overview: emissions, median trend and median of the measurements
}
\keyword{internal}
