% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality.R
\docType{package}
\name{airquality}
\alias{airquality}
\title{This packages allows to retrieve different air quality data from the  OpenWeatherMap API.}
\description{
The package currently has two functions. One to get current airquality data and one to plot the airquality data from the last two weeks.
}
\section{Current airquality}{

\emph{current_aq}
}

\section{Last two weeks airquality}{

\emph{history_aq}
}

