% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{calc_vsp}
\alias{calc_vsp}
\title{Vehicle specific power following Jiménez}
\usage{
calc_vsp(
  speed,
  accel,
  slope,
  vsp.a = 1.1,
  vsp.b = 0.132,
  vsp.c = 0.000302,
  vsp.g = 9.81
)
}
\arguments{
\item{speed}{Speed in m/s.}

\item{accel}{Acceleration in m/s2.}

\item{slope}{Road grade as ratio.}

\item{vsp.a, vsp.b, vsp.c}{Coefficients of the light duty vehicle approximation.}

\item{vsp.g}{Gravitational acceleration in m/s2.}
}
\value{
Vehicle specific power in kW/t.
}
\description{
Vehicle specific power following Jiménez
}
\keyword{internal}
