% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{combine_thresholds}
\alias{combine_thresholds}
\title{Merge air pollution dataset with corresponding threshold limit values dataset}
\usage{
combine_thresholds(data, threshold_values)
}
\description{
Merge air pollution dataset with corresponding threshold limit values dataset
}
\keyword{internal}
