% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert between common units}
\usage{
convert_units(x, in_unit, out_unit, y = NULL)
}
\arguments{
\item{x}{Vector of values to be converted.}

\item{in_unit}{A single character value indicating the units of `x`.}

\item{out_unit}{A single character value indicating the units to convert `x` to.}

\item{y}{(Optional) [UNDER DEVELOPMENT] Vector of extra values required for certain conversions. (i.e. temperature for converting RH to DEWPOINT)}
}
\value{
a vector of `x` converted from `in_unit` to `out_unit`
}
\description{
`convert_units` provides a simple way to convert between common units in air quality / meteorology.
}
\examples{
convert_units(c(-20:20), in_unit = "C", out_unit = "F")
convert_units(c(0:10), in_unit = "PPM", out_unit = "PPB")
}
\seealso{
Other Utilities: 
\code{\link{get_timezone}()},
\code{\link{on_error}()},
\code{\link{swap}()}
}
\concept{Utilities}
