% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abgov.R
\name{get_abgov_data}
\alias{get_abgov_data}
\title{Download air quality station observations from the Alberta (Canada) Government}
\usage{
get_abgov_data(stations, date_range, raw = FALSE, verbose = TRUE)
}
\arguments{
\item{stations}{A character vector of one or more station names to try and get data desired for (see [get_abgov_stations()]).}

\item{date_range}{A datetime vector (or a character vector with UTC dates in "YYYY-MM-DD HH" format) with either 1 or 2 values.
Providing a single value will return data for that hour only,
whereas two values will return data between (and including) those times.
Dates are "backward-looking", so a value of "2019-01-01 01:00" covers from "2019-01-01 00:01"- "2019-01-01 01:00".}

\item{raw}{(Optional) A single logical (TRUE or FALSE) value indicating
if raw data files desired (i.e. without a standardized format). Default is FALSE.}

\item{verbose}{(Optional) A single logical (TRUE or FALSE) value indicating if
non-critical messages/warnings should be printed}
}
\value{
A tibble of hourly observation data for desired station(s) and date range where available.
Columns returned will vary depending on available data from station(s).

Dates UTC time and are "backward-looking", so a value of "2019-01-01 01:00" covers from "2019-01-01 00:01"- "2019-01-01 01:00".
}
\description{
Air pollution monitoring in Canada is done by individual Provinces/Territories,
primarily as a part of the federal National Air Pollution Surveillance (NAPS) program.
The Province of Alberta hosts it's hourly air quality observations
through a public API, providing both historic and real-time raw data.

[get_abgov_data()] provides an easy way to retrieve these observations using
station name(s) (see [get_abgov_stations()]) and a specified date or date range.
}
\examples{
\donttest{
get_abgov_data("Calgary Southeast", c("2024-01-05 00", "2024-01-05 23"))
}
}
\seealso{
[get_abgov_stations()]

Other Data Collection: 
\code{\link{get_abgov_stations}()},
\code{\link{get_airnow_data}()},
\code{\link{get_airnow_stations}()},
\code{\link{get_bcgov_data}()},
\code{\link{get_bcgov_stations}()},
\code{\link{purpleair_api}()}

Other Canadian Air Quality: 
\code{\link{CAAQS}()},
\code{\link{get_abgov_stations}()},
\code{\link{get_bcgov_data}()},
\code{\link{get_bcgov_stations}()},
\code{\link{purpleair_api}()}
}
\concept{Canadian Air Quality}
\concept{Data Collection}
