% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airnow_stations.R
\name{get_airnow_stations}
\alias{get_airnow_stations}
\title{Download air quality station metadata from the US EPA "AirNow" platform}
\usage{
get_airnow_stations(
  date_range = "now",
  time_step = "1 days",
  use_sf = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{date_range}{(Optional).
A datetime vector (or a character vector with dates in "YYYY-MM-DD HH:MM:SS" format, or "now" for current hour) with either 1 or 2 values.
Providing a single value will return data for that hour only,
whereas two values will return data between (and including) those times.
Dates are "backward-looking", so a value of "2019-01-01 01:00" covers from "2019-01-01 00:01"- "2019-01-01 01:00".
Default is "now" (the current hour).}

\item{time_step}{(Optional).
A character string specifying the time step to check data for within `date_range`.
Default is "1 days" (i.e. check every day within `date_range` - may be slow for long date ranges).}

\item{use_sf}{(Optional) a single logical (TRUE/FALSE) value indicating whether or not to return a spatial object. using the `sf` package}

\item{quiet}{(Optional).
A single logical (TRUE or FALSE) value indicating if
non-critical messages/warnings should be silenced.
Default is FALSE.}
}
\value{
A tibble of metadata for the air quality monitoring stations on AirNow.
}
\description{
AirNow is a US EPA nationwide voluntary program which hosts non-validated air quality
observation data from stations in the US and many other countries globally.

The AirNow API provides access to daily metadata files for the available stations at that time.

[get_airnow_stations] provides an easy way to retrieve this metadata (typically to determine station id's to pass to [get_airnow_data])
}
\examples{
\donttest{
# Normal usage
get_airnow_stations()
# if spatial object required
get_airnow_stations(use_sf = TRUE)
# if data for some time in the past required
get_airnow_stations(date_range = lubridate::ymd("2022-01-01"))
# Or a range of time
get_airnow_stations(date_range = lubridate::ymd(c("2022-01-01","2022-01-05")))
}
}
\seealso{
[get_airnow_data]

Other Data Collection: 
\code{\link{get_abgov_data}()},
\code{\link{get_abgov_stations}()},
\code{\link{get_airnow_data}()},
\code{\link{get_bcgov_data}()},
\code{\link{get_bcgov_stations}()},
\code{\link{purpleair_api}()}
}
\concept{Data Collection}
