% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection.R
\name{get_bc_data}
\alias{get_bc_data}
\title{Gather air quality station observations from the British Columbia (Canada) Government}
\usage{
get_bc_data(stations, date_range, raw = FALSE)
}
\arguments{
\item{stations}{A character vector of one or more station IDs (BC EMS IDs) identifying stations data desired for. See also: get_bc_stations()}

\item{date_range}{A datetime vector with two values indicating the start and end dates of desired data window. Will be converted to PST.}

\item{raw}{(Optional) A single logical (TRUE or FALSE) value indicating if raw data files desired (i.e. without standardized column names). Default is FALSE.}
}
\value{
A tibble of hourly observation data for desired (a) station(s) and date range where available.
Columns available will vary depending on available data from station(s).
}
\description{
A short description...
}
\examples{
# For a single station
station = "0450307" # EMS IDs - see get_bc_stations()
# For the years 2019 and 2020
date_range = lubridate::ymd_h(c("2019-01-01 00", "2020-12-31 23"), tz = "Etc/GMT+8")
get_bc_data(station, date_range)

# For multiple stations
stations = c("0450307", "E206898") # EMS IDs - see get_bc_stations()
# For January 2019
date_range = lubridate::ymd_h(c("2019-01-01 00", "2019-01-31 23"), tz = "Etc/GMT+8")
get_bc_data(stations, date_range)
}
\seealso{
[get_bc_stations()]

Other Canadian Air Quality: 
\code{\link{AQHI}()},
\code{\link{AQHI_plus}()},
\code{\link{CAAQS}()}
}
\concept{Canadian Air Quality}
\concept{Data Collection}
