% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bcgov.R
\name{get_bcgov_stations}
\alias{get_bcgov_stations}
\title{Download air quality station metadata from the British Columbia (Canada) Government}
\usage{
get_bcgov_stations(
  years = lubridate::year(lubridate::with_tz(Sys.time(), bcgov_tzone)),
  use_sf = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{years}{(Optional) one or more integer values indicating the year(s) to get metadata for.
Default is the current year.}

\item{use_sf}{(Optional) a single logical (TRUE/FALSE) value indicating whether or not to return a spatial object. using the `sf` package}

\item{quiet}{(Optional) a single logical (TRUE/FALSE) value indicating whether or not to suppress non-critical messages.
Default is FALSE}
}
\value{
A tibble of metadata for British Columbia air quality monitoring stations.
}
\description{
Air pollution monitoring in Canada is done by individual Provinces/Territories,
primarily as a part of the federal National Air Pollution Surveillance (NAPS) program.
The Province of British Columbia hosts it's air quality metadata
through a public FTP site.

[get_bcgov_stations()] provides an easy way to retrieve this metadata (typically to determine station id's to pass to `get_bcgov_data()`)
}
\examples{
\donttest{
# Normal usage
get_bcgov_stations()
# if spatial object required
get_bcgov_stations(use_sf = TRUE)
# if data for past/specific years required
get_bcgov_stations(years = 1998:2000)
}
}
\seealso{
[get_bcgov_data()]

Other Data Collection: 
\code{\link{get_abgov_data}()},
\code{\link{get_abgov_stations}()},
\code{\link{get_airnow_data}()},
\code{\link{get_airnow_stations}()},
\code{\link{get_bcgov_data}()},
\code{\link{purpleair_api}()}

Other Canadian Air Quality: 
\code{\link{CAAQS}()},
\code{\link{get_abgov_data}()},
\code{\link{get_abgov_stations}()},
\code{\link{get_bcgov_data}()},
\code{\link{purpleair_api}()}
}
\concept{Canadian Air Quality}
\concept{Data Collection}
