% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot}
\alias{get_plot}
\title{Get one plot from a plot catalog}
\usage{
get_plot(catalog, plot, parameter = NULL, year = NULL)
}
\arguments{
\item{catalog}{Plot catalog as built by \code{\link[=plot_catalog]{plot_catalog()}}.}

\item{plot}{Name of the plot.}

\item{parameter, year}{Parameter and year of the plot; \code{NULL} if the plot has none.}
}
\value{
A ggplot. Stops with an error of class \code{airquality_plot_error} unless exactly one plot matches.
}
\description{
Get one plot from a plot catalog
}
\keyword{internal}
