% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection.R
\name{get_station_data}
\alias{get_station_data}
\title{Gather air quality observations from multiple data sources}
\usage{
get_station_data(
  locations,
  date_range,
  buffer_dist = 10,
  networks = "all",
  sources = "all"
)
}
\arguments{
\item{locations}{A character vector with at least one value that indicates a
location on Open Street Map that data is desired for (ie. "Prince George, BC, Canada"),
OR an sf object with polygon(s) indicating area of interest.}

\item{date_range}{A datetime vector (or a character vector with UTC dates in "YYYY-MM-DD HH" format) with either 1 or 2 values.
Providing a single value will return data for that hour only,
whereas two values will return data between (and including) those times}

\item{buffer_dist}{(Optional) A single numeric value indicating the distance to buffer the station search location by (typically units of km). Default is 10.}

\item{networks}{(Optional) A character vector indicating which monitoring networks to get data for. Default is "all".}

\item{sources}{(Optional) A character vector indicating which data sources to get data from. Default is "all".}
}
\value{
A tibble of hourly observation data for the specified date range and for
stations within a specified location + buffer from specified monitoring networks
and data sources where available.
Columns available will vary depending on available data from station(s) and data sources.
}
\description{
Gather air quality observations from multiple data sources
}
\examples{
# Get data for all stations within 10 km of Fort St. John, BC for first 1 hour of Jan 2019
get_station_data("Fort St. John, BC, Canada", "2019-01-01 00")

# Get data for all stations within 25 km of Kentucky, USA or Edmonton, AB, Canada
#  for first 1 hour of Jan 2019
get_station_data(c("Kentucky, USA", "Edmonton, AB, Canada"),
                   "2019-01-01 00", buffer_dist = 25)

# Get data for all FEM stations in Kamloops, BC, Canada from the BC Gov't only
#  for first 1 hour of Jan 2019
get_station_data(c("Kamloops, BC, Canada"), "2019-01-01 00", buffer_dist = 0,
                 networks = "FEM", sources = "BC")
}
