% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_expo_cumulative}
\alias{ggplot_expo_cumulative}
\title{Plot relative cumulative exposition distribution using ggplot2}
\usage{
ggplot_expo_cumulative(
  data,
  x,
  y,
  linewidth = 1,
  xlims = c(0, NA),
  xbreaks = waiver(),
  titlelab = NULL,
  captionlab = NULL,
  xlabel = NULL,
  threshold = list(value = NA, label = NULL, labelsize = 4, linetype = 2, linesize = 1),
  theme = ggplot2::theme_minimal()
)
}
\description{
Plot relative cumulative exposition distribution using ggplot2
}
\keyword{internal}
