% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_timeseries_bars}
\alias{ggplot_timeseries_bars}
\title{Plot bar timeseries for yearly population-weighted mean concentration or health-outcome data using ggplot2}
\usage{
ggplot_timeseries_bars(
  data,
  mapping = ggplot2::aes(x = year, y = population_weighted_mean, fill = scenario),
  ylims = c(NA, NA),
  ybreaks = waiver(),
  titlelab = NULL,
  captionlab = NULL,
  theme = ggplot2::theme_minimal()
)
}
\description{
Plot bar timeseries for yearly population-weighted mean concentration or health-outcome data using ggplot2
}
\keyword{internal}
