% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pars_popmean_timeseries}
\alias{plot_pars_popmean_timeseries}
\title{Wrapper to plot timeseries of population-weighted mean concentration data using ggplot2 providing pollutant-specific list}
\usage{
plot_pars_popmean_timeseries(
  data,
  parameters,
  version = "overall",
  id_subareas = NULL,
  y = "population_weighted_mean"
)
}
\description{
Wrapper to plot timeseries of population-weighted mean concentration data using ggplot2 providing pollutant-specific list
}
\keyword{internal}
