% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pars_prelim_deaths_timeseries}
\alias{plot_pars_prelim_deaths_timeseries}
\title{Wrapper to plot timeseries of health-outcome preliminary deaths using ggplot2 providing pollutant-specific list}
\usage{
plot_pars_prelim_deaths_timeseries(data, parameters, relative = FALSE)
}
\description{
Wrapper to plot timeseries of health-outcome preliminary deaths using ggplot2 providing pollutant-specific list
}
\keyword{internal}
