% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotlist_to_tibble}
\alias{plotlist_to_tibble}
\title{Restructures list of plots to a tibble including plots}
\usage{
plotlist_to_tibble(plotlist, type, source)
}
\description{
Restructures list of plots to a tibble including plots
}
\keyword{internal}
