% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{prep_vehicle_specific_power}
\alias{prep_vehicle_specific_power}
\title{Add calculated vehicle specific power to RSD dataset}
\usage{
prep_vehicle_specific_power(data)
}
\description{
Add calculated vehicle specific power to RSD dataset
}
\keyword{internal}
