% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prepare_ressources}
\alias{prepare_ressources}
\title{Prepare data from ressources.csv for use in scripts}
\usage{
prepare_ressources(ressources)
}
\description{
Prepare data from ressources.csv for use in scripts
}
