% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_rsd}
\alias{prepare_rsd}
\title{Merge RSD metadate with dataset, apply filters, calculate vehicle specific power, calculate NOx-emissions and prepare dataset for aggregation}
\usage{
prepare_rsd(data, rsd_auxiliary)
}
\description{
Merge RSD metadate with dataset, apply filters, calculate vehicle specific power, calculate NOx-emissions and prepare dataset for aggregation
}
