% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpleair_data.R
\name{purpleair_api}
\alias{purpleair_api}
\title{Interface with the PurpleAir API}
\usage{
purpleair_api(
  read_key = NULL,
  write_key = NULL,
  channel,
  parameters = NULL,
  verbose = TRUE
)
}
\arguments{
\item{read_key}{(Optional) A single character value of your PurpleAir API read key (see develop.purpleair.com). If NULL, write_key must be provided.}

\item{write_key}{(Optional) A single character value of your PurpleAir API write key (see develop.purpleair.com). If NULL, read_key must be provided.}

\item{channel}{A single character value indicating what API channel to use (Currently supported: keys, organization, sensors, groups)}

\item{parameters}{(Optional) A named list containing paramaters to be supplied to the API request (see api.purpleair.com).}

\item{verbose}{(Optional) A single logical (TRUE or FALSE) value indicating if
non-critical messages/warnings should be printed.}
}
\value{
If channel either "keys" or "organization", or invalid/missing parameters, then a list with the call results
If sensors or groups channel, then a tibble with the data returned from the request
}
\description{
PurpleAir provides access to the observations from their network of low-cost PM2.5 monitors through a points-based API system.
New accounts start with 1M points, and each call has a cost from 1-5 points plus the number of rows times the total column cost
(which depends on the column, typically 1-2 points per column).
Point can be purchased on develop.purpleair.com and cost less if buying in bulk.

For more information, see: api.purpleair.com

In order to retrieve these data you must:
\itemize{
\item make an account on develop.purpleair.com (a google account is required, which can be made using any email)
\item create a new project, click on it after it is created
\item add two API keys, 1 for reading (getting data), 1 for writing (making groups of sensors)
\item (Optional) open your .Renviron for editing (you can use \code{usethis::edit_r_environ("project")} after installing the usethis package)
\item (Optional) add your keys on 2 seperate lines in this format: \code{purpleair_api_read = "YOUR-API-KEY-HERE"}
\item (Optional) refer to keys using \code{read_key = Sys.getenv("purpleair_api_read")} to keep your API key secure
}

\code{purpleair_api()} requires either a read or write key (depending on what you are trying to do), a channel name to make a request to,
and (depending on the request type you want to make) a named list of parameters.
See api.purpleair.com for which parameters are avalailable for each request type.
Based on the provided channel/parameters, the request and estimated points cost will be displayed,
and the request will be made if desired. Set \code{verbose} to \code{FALSE} to silence this and always run the request
(caution, some calls may cost a lot of points)
}
\examples{
\donttest{
read_key = Sys.getenv("purpleair_api_read")
write_key = Sys.getenv("purpleair_api_write")

parameters = list( # see api.purpleair.com for more
  nwlat = 63.595851, nwlng = -135.899856, 
  selat = 63.592657, selng = -135.891057,
  fields = "temperature", 
  sensor_index = 198385, 
  start_timestamp = Sys.time() |> 
    lubridate::with_tz("UTC") - lubridate::minutes(15))


# Get Sensors Data
test = purpleair_api(read_key = read_key, channel = "sensors", 
  parameters = parameters[1:5])

# Get Sensor Data
test = purpleair_api(read_key = read_key, channel = "sensors", 
  parameters = parameters[5:6])

# Get Sensor History
test = purpleair_api(read_key = read_key, channel = "sensors", 
 parameters = parameters[5:7])
}
}
\seealso{
Other Data Collection: 
\code{\link{get_abgov_data}()},
\code{\link{get_abgov_stations}()},
\code{\link{get_airnow_data}()},
\code{\link{get_bcgov_data}()}

Other Canadian Air Quality: 
\code{\link{AQHI}()},
\code{\link{AQHI_plus}()},
\code{\link{CAAQS}()},
\code{\link{get_abgov_data}()},
\code{\link{get_abgov_stations}()},
\code{\link{get_bcgov_data}()}
}
\concept{Canadian Air Quality}
\concept{Data Collection}
\concept{USA Air Quality
#'}
