% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_diagram.R
\name{taylor_diagram}
\alias{taylor_diagram}
\title{Create a Taylor diagram to assess model performance using the relationship between correlation, standard deviation, and centered RMS error.}
\usage{
taylor_diagram(
  dat,
  groups,
  left_cor_limit = NULL,
  mod_colours = "default",
  mod_fills = "default",
  mod_shapes = "default",
  mod_size = 3,
  mod_stroke = 2,
  obs_colour = "purple",
  obs_shape = 16,
  obs_size = 3,
  obs_stroke = 2,
  obs_label = "Observed",
  obs_label_vjust = -1,
  obs_label_hjust = -0.05,
  cor_colour = "lightgrey",
  cor_linetype = "solid",
  rmse_colour = "brown",
  rmse_linetype = "dotted",
  rmse_label_pos = 80,
  sd_colour = "black",
  sd_linetypes = c(obs = "dashed", max = "solid", other = "dotted"),
  plot_padding = 2,
  labels_padding = 2
)
}
\value{
A ggplot object of your taylor diagram
}
\description{
Blah Blah Blah
}
\examples{
# Make test data
data = data.frame(
  obs = c(3, 2, 3, 3, 4, 5, 6, 6, 7, 3),
  mod = c(0.1, 0.9, 2.0, 2.9, 3.9, 4.7, 6.1, 7, 7.3, 5.2),
  group = factor(c(rep(1, 5), rep(2, 5)))
)
# Basic usage
taylor_diagram(data, groups = c(Group = "group"))
# Force 0 on left axis
taylor_diagram(data, groups = c(Group = "group"), 
  left_cor_limit = 0, rmse_label_pos = 130)
# Change colours / shapes
taylor_diagram(data, groups = c(Group = "group"), 
  mod_colours = c("AB" = "pink", "BC" = "blue"), 
  mod_fills = c("EGG" = "white", "PA" = "darkgrey"),
  mod_shapes = c("FALSE" = 23, "TRUE" = 22),
  mod_size = 4, mod_stroke = 6,
  obs_colour = "brown", obs_shape = 23, obs_size = 6, 
  cor_colour = "orange", cor_linetype = "dotdash",
  rmse_colour = "green", rmse_linetype = "longdash",
  sd_colour = "purple", sd_linetypes = c(obs = "solid", max = "dotted", other = "dashed")
  )
# Adjust text positioning
taylor_diagram(data, groups = c(Group = "group"),
  plot_padding = 4, labels_padding = 1, rmse_label_pos = 80)

# Save plot
# ggplot2::ggsave("test.png", width = 7, height = 5, units = "in")
}
\concept{Data Visualisation}
\concept{Model Validation}
