% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_diagram.R
\name{taylor_diagram}
\alias{taylor_diagram}
\title{Create a Taylor diagram to assess model performance using the relationship between correlation, standard deviation, and centered RMS error.}
\usage{
taylor_diagram(
  dat,
  data_cols = c(obs = "obs", mod = "mod"),
  groups,
  left_cor_limit = NULL,
  right_sd_limit = NULL,
  mod_colours = "default",
  mod_fills = "default",
  mod_shapes = "default",
  mod_size = 3,
  mod_stroke = 2,
  obs_colour = "purple",
  obs_shape = 16,
  obs_size = 3,
  obs_stroke = 2,
  obs_label = "Observed",
  obs_label_vjust = -1,
  obs_label_hjust = -0.05,
  cor_colour = "lightgrey",
  cor_linetype = "solid",
  rmse_colour = "brown",
  rmse_linetype = "dotted",
  rmse_label_pos = 80,
  sd_colour = "black",
  sd_linetypes = c(obs = "dashed", max = "solid", other = "dotted"),
  plot_padding = 2,
  labels_padding = 2
)
}
\arguments{
\item{dat}{Paired observation and model data with (at least) all columns in `data_cols` and `groups`.}

\item{data_cols}{(Optional) a character vector with 2 values indication column names in `dat` to get observed and modelled values. Must have names `"obs","mod"`.}

\item{groups}{a character vector with between 1 and 3 column names to use as groups. The first value will be used for `colour`, the second (if present) will be used for `shape`, and the third (if present) will be used for `fill` when adding model data points.}

\item{left_cor_limit}{(Optional) a single numeric value indicating the minimum correlation value to display (from -1 to +1). If not provided, the nearest 0.1 below the minimum correlation in the data will be used.}

\item{right_sd_limit}{(Optional) a single numeric value indicating the maximum standard deviation value to display (>=0). If not provided, the nearest "pretty" above the maximum standard deviation in the data will be used.}

\item{mod_colours, mod_shapes, mod_fills}{(Optional) a named vector of colours/shapes to use for the provided `groups` where the names correspond to values in that group column to assign each colour/shape to (i.e `c("group_1" = "red", ...)`)}

\item{mod_size, mod_stroke}{(Optional) single numeric value indicating the size/stroke of the model data points}

\item{obs_colour, obs_shape, obs_size, obs_stroke}{(Optional) a single value indicating the colour/shape/size/stroke of the observed data point}

\item{obs_label}{(Optional) a single character value indicating the text to displat for the observed point.}

\item{obs_label_vjust, obs_label_hjust}{(Optional) a single numeric value indicating how to position the observed label relative to the observed point.}

\item{cor_colour, cor_linetype}{(Optional) a single value indicating the colour/linetype of the correlation grid lines.}

\item{rmse_colour, rmse_linetype}{(Optional) a single value indicating the colour/linetype of the rmse circles originating from the observed point.}

\item{rmse_label_pos}{(Optional) a single value (0-360)indicating the location of the labels for the rmse circles.}

\item{sd_colour}{(Optional) a single value indicating the colour of the standard deviation arcs.}

\item{sd_linetypes}{(Optional) a character vector with 3 line types and names `"obs", "max", "other"` indicating the line types of standard deviation arcs.}

\item{plot_padding, labels_padding}{(Optional) a single numeric value indicating how much spacing (standard deviation units) to add to most text labels.}
}
\value{
A ggplot object of your taylor diagram.
}
\description{
Blah Blah Blah Taylor (2001) Blah Blah Blah
}
\examples{
\dontrun{
# Make test data
data = as.data.frame(datasets::ChickWeight) |>
  dplyr::filter(.data$Chick == 1) |>
  tidyr::pivot_wider(names_from = "Chick", values_from = "weight") |>
  dplyr::full_join(
    as.data.frame(datasets::ChickWeight) |>
      dplyr::filter(.data$Chick != 1)
  ) |> 
  dplyr::rename(obs = `1`, mod = "weight") |>
  dplyr::mutate(Chick = factor(round(as.numeric(.data$Chick) / 20)))
# Basic usage
taylor_diagram(data, groups = c(Diet = "Diet", Chick = "Chick"))
# Force 0 on left axis
taylor_diagram(data, groups = c(Group = "group"), 
  left_cor_limit = 0, rmse_label_pos = 130)
# Change colours / shapes
taylor_diagram(data, groups = c(Group = "group"), 
  mod_colours = c("AB" = "pink", "BC" = "blue"), 
  mod_fills = c("EGG" = "white", "PA" = "darkgrey"),
  mod_shapes = c("FALSE" = 23, "TRUE" = 22),
  mod_size = 4, mod_stroke = 6,
  obs_colour = "brown", obs_shape = 23, obs_size = 6, 
  cor_colour = "orange", cor_linetype = "dotdash",
  rmse_colour = "green", rmse_linetype = "longdash",
  sd_colour = "purple", sd_linetypes = c(obs = "solid", max = "dotted", other = "dashed")
  )
# Adjust text positioning
taylor_diagram(data, groups = c(Group = "group"),
  plot_padding = 4, labels_padding = 1, rmse_label_pos = 80)

# Save plot
# ggplot2::ggsave("test.png", width = 7, height = 5, units = "in")
}
}
\concept{Data Visualisation}
\concept{Model Validation}
